expense-list: category;

+++ src/main/java/com/arobs/repository/ExpenseCategoryRepository.java
@@ -10,7 +10,7 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface ExpenseCategoryRepository extends JpaRepository<ExpenseCategory, Integer> {
+public interface ExpenseCategoryRepository extends JpaRepository<ExpenseCategory, Long> {
 
     @Query("SELECT ec FROM ExpenseCategory ec WHERE ec.tenantId IS NULL")
     List<ExpenseCategory> findDefault();
